backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
numpy>=1.21.0
pandas>=1.3.0
pydantic>=1.8.0
motor>=2.5.0
redis>=4.0.0
PyJWT>=2.3.0
python-multipart>=0.0.5
pytest>=6.2.0
pytest-asyncio>=0.15.0
openai>=1.0.0
pymongo>=3.12.0
# sui-sdk==0.1.1.dev0  # For Sui blockchain integration (temporarily disabled due to installation issues)